[fix] Post新規作成ページのデザインを変更。
